requirements.txt
beautifulsoup4>=4.8.0
icalendar>=4.0.3
requests>=2.22.0
PyGithub>=1.44.1
PyYAML>=5.1
tabulate>=0.8.5
lxml>=4.4.1
